rename some archives at home
