+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".view.HomeActivity"/>
+        <activity android:name=".presentation.activity.HomeActivity"/>
     </application>
 
 </manifest>
