+++ app/src/main/java/wottrich/github/io/androidsmartchecklist/di/HomeModule.kt
@@ -1,9 +1,9 @@
-package wottrich.github.io.androidsmartchecklist.injection
+package wottrich.github.io.androidsmartchecklist.di
 
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
-import wottrich.github.io.androidsmartchecklist.view.DrawerViewModel
-import wottrich.github.io.androidsmartchecklist.view.HomeViewModel
+import wottrich.github.io.androidsmartchecklist.presentation.viewmodel.DrawerViewModel
+import wottrich.github.io.androidsmartchecklist.presentation.viewmodel.HomeViewModel
 
 /**
  * @author Wottrich
