+++ app/src/main/java/wottrich/github/io/androidsmartchecklist/presentation/ui/HomeDrawerChecklistItemComponent.kt
@@ -1,4 +1,4 @@
-package wottrich.github.io.androidsmartchecklist.ui
+package wottrich.github.io.androidsmartchecklist.presentation.ui
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -7,6 +7,7 @@ import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.composed
 import androidx.compose.ui.unit.dp
 import wottrich.github.io.baseui.ui.Dimens.BaseFour
 import wottrich.github.io.baseui.ui.ListItem
@@ -25,9 +26,10 @@ import wottrich.github.io.database.entity.Checklist
  */
 
 private val ChecklistItemBorderStroke = 1.dp
+private val ChecklistItemShape = RoundedCornerShape(BaseFour.SizeTwo)
 
 @Composable
-fun ChecklistItem(
+fun HomeDrawerChecklistItemComponent(
     checklist: Checklist,
     onItemClick: () -> Unit
 ) {
@@ -45,25 +47,11 @@ private fun ItemContent(
     checklist: Checklist
 ) {
 
-    val shape = RoundedCornerShape(BaseFour.SizeTwo)
-
-    val paddingModifier = Modifier.padding(
-        top = BaseFour.SizeTwo,
-        start = BaseFour.SizeTwo,
-        end = BaseFour.SizeTwo
-    )
-
-    val surfaceModifier = if (checklist.isSelected) {
-        paddingModifier.border(
-            width = ChecklistItemBorderStroke,
-            color = SmartChecklistTheme.colors.onPrimary,
-            shape = shape
-        )
-    } else paddingModifier
+    val surfaceModifier = Modifier.checklistItemPadding().addBorderIfSelected(checklist.isSelected)
 
     Surface(
         modifier = surfaceModifier,
-        shape = shape,
+        shape = ChecklistItemShape,
         elevation = BaseFour.SizeOne,
         color = SmartChecklistTheme.colors.onSurface
     ) {
@@ -78,3 +66,21 @@ private fun ItemContent(
         )
     }
 }
+
+private fun Modifier.addBorderIfSelected(isSelected: Boolean): Modifier =
+    composed {
+        if (isSelected) {
+            this.border(
+                width = ChecklistItemBorderStroke,
+                color = SmartChecklistTheme.colors.onPrimary,
+                shape = ChecklistItemShape
+            )
+        } else this
+    }
+
+fun Modifier.checklistItemPadding(): Modifier =
+    composed { Modifier.padding(
+        top = BaseFour.SizeTwo,
+        start = BaseFour.SizeTwo,
+        end = BaseFour.SizeTwo
+    ) }
